requirements.txt
quart
quart-cors
sqlalchemy
requests
python-dotenv
hypercorn
tqdm